bitfield_gen: allow non-contiguous tag fields
A tagged union can now optionally use multiple fields to indicate the
tag. These are called "sliced" tags in the code. The tag fields have to
be at the same position and width in each block of the tagged unions,
and all tag fields have to be within the same word.

The enum value of the overall tag is the plain unshifted number that
comes out when the tag is extracted by masking, e.g. if the tag
consists of two fields of width 1 at positions 0 and 2, the possible
enum values are 0b000, 0b001, 0b100, and 0b101. This only applies when
tag slices are used.

When non-contiguous (=sliced) tags are used, tag class masks cannot be
used at the same time for this tagged union.

This commit extends the parser and code generator, but not yet the proof
generator for the new feature.

Example usage:

```
tagged_union ex exType(x,y) {
    tag ex1 0x00002
    tag ex2 0x10003
}
```

where `x` and `y` are tag fields (declared in blocks not shown here
with width 2 and 1 and positions 0 and 16).

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>