bitfield_gen: mark unsupported proofs with oops
These were previously marked with `sorry` which is dangerous, because
it means that the prover will accept the lemma without proof. This will
be checked in final proofs, so it is not soundness critical, but it
might go unnoticed for a while during development.

Replacing with `oops` means there are no unproven lemmas in the output
unless specifically requested with `--sorry_lemmas`, which can also be
used for development when unsupported combinations are triggered.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>